Change atomics flags for Morello
Before this patch we had disabled both `-moutline-atomics` and
`-fno-inline-atomics` for purecap Morello.  There is no reason to
disable `-moutline-atomics` for Morello.  The purpose of
`-moutline-atomics` is to create a binary which will run correctly on
Armv8-a but take advantage of the LSE atomics when running on an
architecture which has those.

Since the Morello architecture has the LSE atomic instructions for
non-capability data types, we naturally fall into the same clauses as
non-capability architectures which have the LSE instructions -- avoiding
the outline atomics implementations all together.
We do have to add a clause which avoids using the outline-atomic stubs
for capability data types.  For -mfake-capability it's worth using these
since the actual data type at runtime is still a long unsigned integer.
For Morello there is no point using the outline stubs since there is no
existing "faster" version to choose at runtime.

In other words, the point of these outline stubs is to allow using a
faster version of an atomic operation when running on an architecture
which provides such a faster version, while still producing a binary
which can run on architectures which do not have this faster version.
As yet there is no different architecture which would be concerned with
atomic operations on capabilities, hence there is no point in worrying
about the out-of-line stubs.

W.r.t. `-fno-inline-atomics`: Morello has inline implementations of all
the atomic operations that libatomic handles (and in fact also has an
inline implementation of the capability operations which libatomic does
not handle).  Because of that, libatomic is compiled just fine and
the only thing we need to do in order to support `-fno-inline-atomics`
on everything except capability data is to remove the error on receiving
the command line flag.
Doing so means that all the existing libatomic testcases pass.

This does mean that we're in a situation where when compiling with
`-fno-inline-atomics` and trying to perform atomic operations on
capability data we get a linker error rather than a nice compile-time
error.
However, as Richard S pointed out, `-fno-inline-atomics` is not a
documented flag and it hence seems reasonable to claim that it is not
supported.

One alternate route would be to keep the current compile-time error
message when attempting to use `-fno-inline-atomics` with Morello.
I feel that it is worth allowing for those data types that it
automatically works on.

Another alternate route would be to add capability data type operations
into libatomic so that the feature is fully supported.
This is the eventual goal, but is left for the moment in order to focus
on higher priority things.