Use effective-target for hybrid tests
Many of our hybrid-specific tests use dg-skip-if looking for the
`-mabi=purecap` flag to avoid running on purecap targets.  This does not
work hosted since our hosted implementations run purecap by default.

Here we introduce and use a new effective-target
`cheri_capability_hybrid` to make the selection more robust.

Some precise-bounds-padding tests were using a TLS variable.  These did
not have an associated `dg-warning` to catch the warning that we emit
when the target actually has native TLS support.  This patch adds said
dg-warning so that we do not get an "excess errors" testsuite failure
when running these on the board.