Testism adjustments for pure capability
array25.C
This testcase uses a global buffer as its place to allocate a structure.
During initialisation on unoptimised code, the structure initialises its
sub-structures with zeros.  Since the sub-structure contains a pointer
that includes storing a capability.  Said global buffer is not aligned.
Hence we get a BUS error.

Here we ensure that the alignment is specified for pure capability
targets.

struct-layout-1.exp
This testcase generates multiple different structures with different
alignment behaviours.  It purposefully generates misaligned pointers.
These cause failures on purecap.
This is a G++ variation of the same problem that we had in GCC, and we
simply avoid the tests in the same way that we avoided them in GCC.

g++.abi/vtable2.C
This testcase uses ptrdiff_t for vtable entries.  On purecap we have
capability-sized and capability-aligned vtable entries.  There seems to
be at least one existing target architecture (see gcc/config/darwin.h)
which uses different types for ptrdiff_t and intptr_t (although I
haven't checked whether these types boil down to the same size data on
hardware), so rather than change the testcase for existing architectures
we only change the type for purecap.