c, cp: CHERI warning improvements
Currently for C code like:

char *f(long x) { return (char *)x; }

we emit the warning "cast from provenance-free integer type to pointer
type will give pointer that can not be dereferenced" but we reuse the
-Wint-to-pointer-cast warning flag for this.

This patch adds the -Wcheri-capability-misuse flag to align better with
CHERI LLVM and switches the C frontend to using that flag for this
warning. We also implement the warning in the C++ frontend, and add
tests for this warning for both C and C++.

We also take this opportunity to implement the -Wcheri warning group,
and add all existing CHERI warnings under this flag. This enables using
-Wno-cheri to turn all the CHERI warnings off, and e.g. -Werror=cheri to
turn all CHERI warnings into errors.