testsuite: Fixes for -Wcheri-capability-misuse
This patch adjusts the testsuite for the following recent changes:
 - -Wcheri-capability-misuse has been pulled out as a separate warning
   from -Wint-to-pointer-cast
 - The C++ frontend now implements -Wcheri-capability-misuse (now at
   parity with the C frontend).

We also take the opportunity to fix up a pre-existing issue whereby all
the dg-warning directives looking for this warning were gated on
aarch64_capability_any, which includes hybrid, but of course these
warnings aren't expected to be triggered on hybrid.

To fix this issue we introduce a new effective-target
"cheri_pointers_are_caps", which is enabled for purecap and
fake-capability Morello but not for hybrid CHERI. We then use this to
gate the dg-warning directives.