Update libstdc++ `pool_allocator` for capabilities
The existing pool_allocator implementation does not work with
capabilities.  The problem we see is a SIGBUS on an attempt to store a
capability to an unaligned location.

This problem was due to the mechanism by which a free list is maintained
in this allocator.  We store a pointer to the "head" of a linked list
of free allocations in an array indexed by size of allocation.  During
allocation the existing implementation ensured that the alignment of
each chunk given out was at least 8 bytes (and that the size of each
chunk given out was at least 8 bytes).  This meant that the alignment
and size of each chunk was large enough for a pointer.

With the larger sizes and higher alignment requirements of pointers on
capability architectures this was no longer enough.  Here we change
`_S_align` (which is used to both maintain alignment and round up
allocations) to `__SIZEOF_POINTER__` when that macro exists.  This both
provides the necessary alignment and size for the larger capabilities.

There is a maximum size of 128 bytes beyond which this allocator falls
back to `new` and `delete`.  This size is smaller than the size at which
padding and extra alignment become necessary to ensure that CHERI bounds
compression do not allow access beyond the current object that we're
allocating.  Hence in updating this allocator we do not need to concern
ourselves with such padding and extra alignment.

----
In order to bound the pointers returned by this allocator we can simply
add a call to a builtin function applying the necessary bounds at the
end of the `allocate` member function.
This works well apart from one complication due to the free list
mechanism described above.  During deallocation the original pointer is
given back to the allocator, and the allocator uses this pointer to
store a link to the "next" free chunk while updating the "head" of its
free list for this size to be the given pointer.  If the bounds on this
pointer are not large enough to span the size of a capability, then we
can not store a capability in this position (even if the space we had
carved out for this chunk was padded up to the size of a capability).
A similar problem arises once an allocation has been provided, reclaimed
onto a free list, and is then given out again.  Given that allocations
are made and categorized based on multiples of `_S_align`, if we bound
the size of an allocation to the number of bytes it needs then it may
not span the entire allocation hunk that was given out.  Hence after
this allocation has been reclaimed onto the free list it may be given
out to another consumer which needs the entirety of the allocation yet
is not given a pointer with wide enough bounds.

In order to apply the most restrictive bounds possible on the pointers
given out, we would need to maintain another data structure containing
wider-bounds capabilities from which to extend the necessary bounds of
the pointers given to `deallocate`.

Rather than add such extra complexity, here we choose to give out
pointers with bounds that are always rounded up with `_M_round_up` to
the same size that we associate allocations in our free list.  This does
mean that programming errors which read slightly past the end
allocations are not caught, but it still restricts the permissions given
to the returned pointer enough so that it can not access any
neighbouring object.  This is considered a reasonable security tradeoff.

----
One extra feature that was considered was that due to the aforementioned
mechanism by which the free list is maintained, each allocation would
usually be returned containing a capability that can access the next
free allocation.
This indirectly provides more permissions than are necessary to the user
of this API.  Those extra permissions could theoretically be used to put
some other memory region on the free list (triggering the allocator to
write to some memory that it should not).  The only memory a user could
cause to be accessed would be memory that user already had access to
(since they would need a valid capability), and it seems relatively
unlikely to accidentally affect the free list when compared to the
likelihood of accidentally writing past the end of an allocation.
Nevertheless it seems reasonable to clear this capability on allocation,
and we perform that by a simple write to clear the free list link when
returning a pointer from the free list.

----
Of note is that there is nothing in pool_allocator that handles increasing
the alignment of a pool from the original pointer provided by
`::operator new` in `_M_allocate_chunk`.  This means that increasing the
`_S_align` member beyond 16 bytes (which is the allocation alignment of
`malloc`) would not continue to increase the alignment of the pointers this
allocator gives out.  Luckily that doesn't matter for us, but it's worth noting
that we do rely on the alignment of pointers that come back from `::operator
new`.