Update mt_allocator for CHERI
The implementation basics this allocator are as follows:
  - There are N "bins" for each power-of-two size.  N is determined
    based on the maximum size beyond which we do use `new` directly.
  - Each bin contains a linked list of "chunks" that have been
    allocated.  This list is only used at cleanup-time so that we can
    call `delete` on each one.
  - Each bin then contains a linked-list of "records" that represent
    available allocations of the correct size.
  - When there are no available "records" we generate some more by
    obtaining an `_M_chunk_size` block from `new` and splitting it up
    into `(_M_chunk_size - address metadata) / sizeof(record)` records.

----
The existing errors that we see in our testsuite were all to do with
misalignment of capabilities in our allocations.  This was happening
when we were attempting to store our linked-list pointers in misaligned
locations.  There is an `_M_align` configuration member which determines
the alignment required for a record.  This must also be `>=
sizeof(_Block_record)` since the alignment also corresponds to the size
left for each record metadata.

The existing alignment was 8.  We update it to use the larger of 8 and
`__SIZEOF_POINTER__` where possible given that the size of a
`_Block_record` is a pointer.  This avoids all alignment problems that
we see in the testcase.

Considering if extra padding and alignment is necessary for large
allocations, we see that the maximum allocation size before falling back
to `new` is configurable.  However, as the comments describe this has a
maximum of 32768.  It happens that the alignment requirement to allow
CHERI to precisely bound an allocation of 32768 bytes is 16, which is
the same alignment requirement we already have in order to store
capabilities in the metadata.  Hence we do not have to worry about
padding and alignment for CHERI bounds.

----
Given that the metadata recorded in these allocations contains a pointer
that would give the user more permissions (especially around permissions
to access other objects) than are necessary, we zero out this metadata
area before returning it to the user.  This information does not need to
be maintained over the period that the user has access to it.

----
Restricting the bounds of returned allocations is more tricky.  The act
of applying bounds is relatively simple -- we have to call the
appropriate bounding function everywhere that a "record" is made.  The
problem occurs during deallocation when the allocation given out must be
reclaimed by adding metadata to the space.  In mt_alloc the metadata is
stored just before the allocation returned to the user.  This stands in
contrast to the pool_allocator which stores metadata at the start of the
allocation.  This means that a basic implementation of narrowing bounds
would need to include 16 bytes before the allocation returned as well as
the total allocation returned.

On top of this, since this allocator is a power-of-two allocator (as
opposed to a multiple-of-alignment sized allocator like pool_allocator)
the space at the end of an allocation that would need to be accessible
in order for any reclaimed allocations to have bounds spanning the
entire range they need may be quite large.  E.g. an allocation of 8193
bytes would be satisfied from the free list of blocks of 16384 (2^14)
bytes.  Without any mechanism to widen the bounds of a deallocated
pointer we would have to give the user a capability with bounds spanning
the entire 16384 bytes.

This implies that the trade-off in forgoing tight bounds in order to not
change the implementation of this allocator is much more dramatic than
the corresponding trade-off in pool_allocator.  However the decision is
still in much the same vein -- tight-bounds requires some extra data
structure to maintain extra permissions for use in increasing bounds of
pointers that are being deallocated, while lax bounds miss more
out-of-bounds accesses but still do not provide access to any other
objects that the user did not have access to before (only provides
access to padding at the end of an allocation).

Based on the fact that this allocator keeps a record of allocations that
it has made, these allocations could be used to widen the bounds of any
block returned to the allocator by the user.  However this would require
either a new data structure to look these things up (e.g. hash table) or
a linked-list traverse on the existing data structure.

Given this particular allocator does not seem to be used in many places
(a debian code search for `__mt_alloc` only found few packages outside
of compilers), we have decided to forego adjusting this particular
allocator to have the tightest bounds available.