morello: Narrowing of outgoing varargs
This patch adds bounds narrowing and permission clearing on the outgoing
varargs pointer (c9) for pure-capability Morello. The bounds narrowing
is needed for conformance with AAPCS64-cap (and is of course desirable
from a security perspective). The permission clearing follows what is
done in Morello LLVM.

Note we don't handle padding and alignment to ensure tight bounds in the
case of > 4096 varargs, but this is not required for ABI conformance. It
could be added in the future as an additional enhancement, if desired.

To assist in implementing the permission clearing, we refactor the code
that defines the __{ARM,CHERI}_CAP_PERMISSION_* macros. Even the CHERI
permission macros should be defined by the backend, since the values are
target specific (they differ between CHERI architectures). Defining them
in the backend also allows us to use a shared enum across the backend
which can then be used for codegen. We make use of that in this patch in
order to clear the write and execute permissions on c9 for outgoing
varargs.

While refactoring the CHERI permsision macros, we stop defining
__CHERI_CAP_PERMISSION_PERMIT_CCALL__, which the ACLE explicitly says
should not be defined on Morello.