Inlining memcpy and tag-preserving behaviour
This commit combines two fundamental changes.  First we are bailing out of
inlining memcpy in another situation.  Second, we are refactoring the
way that we avoid inlining memcpy using `bucket_type_for_size`.

---
Bail out of performing any inlining transformation of memcpy/memmove
when the size of the operation can hold a capability.  Inlining a memcpy
is slightly more complicated in the presence of capabilities.  There are
two cases where we know operations that we can make.

1) When we are moving between two destinations that are known to be
   aligned we can use `ldr`/`str` on capability registers and be certain
   that the operation will maintain the validity of any capability that
   may exist in the source memory region.
2) When we are moving from a source that is known to not contain any
   capabilities we may use integer loads to perform the operation.

A naive implementation (and the one that we had until now) may determine
that a memcpy source must not have capabilities if the types of the
pointers given to the memcpy call do not contain capabilities.  However
this is not necessarily the case.

One case where this is evident is in passing pointers to a union member,
and that is indeed the testcase which we add for this patch.

This patch implements a conservative approach with the idea that later
improvements could identify more complex situations where a memory
operation may be determined to act on known non-capability data.

N.b. for future reference, it may be useful to look at the discussion
below for information on which cases may be identified as operations on
known non-capability data.
https://github.com/CTSRD-CHERI/llvm-project/pull/506

---
A previous commit fixed problems in the C++ frontend inlining a 16 byte
memcpy into integer loads and stores.  We fixed this by avoiding
inlining such calls in the case where the size may contain capabilities.
The C frontend already had accounted for such cases by still inlining
when the alignment was known to be enough for capabilities.

The reason that the C++ and C frontends were not behaving similarly was
due to the `bucket_type_for_size` function using the
`lang_hooks.types.cap_type_for_size` hook.  This had not been set up for
the C++ frontend.

Here we use the same hook implementation for the C++ frontend as the C
frontend and hence can use the same method to avoid problematic memcpy
inlining for both.

N.b. while using this approach we noticed that the existing code can
already handle capability loads and stores in the branch which uses
`bucket_type_for_size`.  The existing checks for alignment are
sufficient to catch the extra requirements on capabilities in a general
way.

The only thing that appeared troublesome was using
`fold_const_aggregate_ref`, but this is logically checking to see if our
value can be determined at compile-time by recognising that it is a
reference into a constant aggregate.  Hence the existing code should
correctly handle capability types.