Add builtin align functions
This patch is implementing builtins for aligning and checking alignment of
pointers and integers. Change supports purecap, hybrid and vanilla.

This change introduces three new user-facing builtins (which work on
both pointers and integers) that can be used instead of common bit-wise
arithmetic:

	__builtin_align_down (x, N)  ->    x & ~(N-1)
	__builtin_align_up (x, N)    ->  ( x +  (N-1) ) & ~(N-1)
	__builtin_is_aligned (x, N)  ->  ( x &  (N-1) ) == 0

These builtins are overloaded over integral parameter types and pointer
types.

In addition, if both builtins parameters are const integers, this change
folds these functions and calculates alignment value.
Since these builtins need to work on capability targets, we need to
provide two implementations.  One for integral types and one for
capability typed values.

The integral typed version is encoded with the user-visible
BUILT_IN_*_INT, and the capability version is encoded with
BUILT_IN_*_CAP.  These are chosen at builtin resolution time.

In order to implement these builtins in Morello we implement the
align_address_up and align_address_down optabs for CADImode operands.

---
Integral parameter types are implemented by all using the same function
which acts on an unsigned long long.  This means that trying to use
these new alignment functions on an __int128 would lose precision.  An
error is emitted on attempting such a thing.

---
We do not implement folding for capability-typed constants.  As it
stands something like `(__intcap_t)30` actually ends up in the folding
routines as a `NOP_EXPR` casting an integral type 128 to an
`INTCAP_TYPE`.  Hence this fold would never trigger.

---
To ensure bad constexpr expressions are given an error we must do two
things.  1) Such expressions must not be fully evaluated when expression
evaluation is speculatively attempted.  2) We should emit an error when
the expression is attempted to be evaluated in a context where full
evaluation is required.

Main Author:
    Przemyslaw Wirkus <przemyslaw.wirkus@arm.com>