Add folding for replace_address_value
This failure to fold a REPLACE_ADDRESS_VALUE internal function call was
causing some test failures in the libstdc++ testsuite when using an
integer casted to a pointer as a parameter in a builtin function call
that the testcase required to get folded away.

The initial call to `fold_build_replace_address_value_loc` is made with
a relatively complex tree as its argument (the "capability value" is
given as the negate_expr of a constant) which is something that we
really don't want to have to handle.

However it seems that in the C++ frontend more folding is attempted
later on.  This folding is attempted using the general function `fold`
defined in fold-const.c.  This is called via `cp_fold`.  If we simply
allow this function to handle the same sort of a REPLACE_ADDRESS_VALUE
on INTEGER_CST arguments that we already handle while building the
internal function call then we observe that the required folding
happens.

N.b. based on some comments in `c_fully_fold_internal`, I believe that
handling replace_address_value in `fold` was not necessary for the C
language since the C language already handled all folding of function
arguments while building the expression up.  Hence this does not appear
to have changed the C frontend behaviour.