Update PMR allocators for CHERI
The allocation methods of the C++17 allocators based of the
memory_resource interface are different to each other.  However it
happens that the two main allocation methods naturally provide the
ability to return precisely bounded pointers from allocation requests.

The two allocator approaches built on top of memory_resource in
libstdc++ are that for the monotonic_resource and that for the
{un,}synchronized_pool_resource.

---
The monotonic_resource is a very simple allocator.  It never reclaims
memory that has been freed by the user.  This allocator maintains a
large buffer (which it gets from `new`).  When an allocation is
requested this allocator moves the "current" pointer along in this
buffer an amount according to the allocation required.  The space left
behind is the space given to the user.

It is clear that returning a tightly bounded capability from such an
allocation would cause no problems.

---
The {un,}synchronized_pool_resource classes are more complicated.  These
both maintain a set of pools from which to allocate, and each pool is
implemented by a vector of `chunk` structures.  These `chunk` structures
allocate blocks by determining which bits in a bitset are free.  An
allocation consists of finding a free bit, returning a pointer generated
from the base pointer to a chunk plus an offset according to which bit
was chosen.  A deallocation consists of finding which bit in which chunk
a pointer corresponded to and clearing that bit.

Since the allocation is always generated from an internal base pointer
and the deallocation only needs the address of a capability, providing a
bounded pointer naturally fits with this mechanism of tracking
allocations.  There is nothing to do beyond ensuring that the returned
pointer is bounded according to the size of allocation.

---
In contrast to the ext/pool_allocator and ext/mt_allocator, these
allocators do allow allocations larger than the threshold at which we
need to pay attention to padding and alignment concerns for CHERI
precise bounds.

To handle that we can use the two builtins introduced for this reason:
  - __builtin_cheri_round_representable_length
  - __builtin_cheri_representable_alignment_mask

Across the allocator code there are a few places which look like they
might need accounting for alignment but do not.  Here we justify the
lack of extra code in each place.

In __pool_resource::_Pool::replenish, we replenish the slab backing
allocations.  This slab is obtained from some underlying allocator
(by default the `new_delete_resource`, but could be anything provided by
the user).  The alignment we request for this slab needs to be enough
for the chunks that we give out (it is relied upon in the allocation
routines of {un,}synchronized_pool_resource).  We can not rely on the
underlying allocator providing us with an aligned pointer, since even if
it is adjusted for CHERI precise bounds it is valid for the returned
pointer to be unaligned with padding and bounds extending to smaller
addresses than the pointer that is returned.  That means that the
alignment we request must be as large as we require.

It happens that the existing algorithm already ensures that the
alignment requested is std::bit_ceil(__block_size) where __block_size is
the size that this chunk will be handling.  Given the properties of
CHERI bounds compression, this alignment is always enough to provide
precise bounds on the given block size.

{un,}synchronized_pool_resource chooses the block_size from which to
satisfy allocations based on the maximum of the length and alignment
requested.  This only relates to the block from which we satisfy
requests, and does not affect the bytes we actually want to provide to
the user.  Hence as long as we have adjusted our request for CHERI
precise bounds before this adjustment there is no problem.

---
There is a static assert in memory_resource.cc on the size of a `struct
chunk`.  This is there to ensure that the padding at the end of a
`struct bitset` is used in the chunk structure that inherits from it.
The increased size and alignment requirements of a pointer on a CHERI
architecture means that there is more padding in a `struct chunk`, and
hence the static_assert fails with the existing expression.  Here we
adjust the assertion expression and include a comment as to where the
padding has come from.

---
We had to update the testcase unsynchronized_pool_resource/allocate.cc
to account for the fact that the size and alignment requested by the
user may be modified by the time the allocation request is passed down
to an underlying allocator.  We hence have to adjust the check according
to CHERI bounds compression adjustments.