c, cp: Switch to handling __capability as a keyword
Prior to this patch, __capability was handled as a built-in macro which
expanded to __attribute__((__cheri_capability__)). This is how Morello
LLVM also used to handle this construct, but LLVM has since switched to
treating __capability as a proper keyword instead. This patch makes the
corresponding change for GCC.

This change is important because keywords / type qualifiers are treated
differently to GNU attributes in the C/C++ grammars, so some parses are
different. E.g. we used to reject code like:

struct {
  int x;
} __capability *p;

since, when treated as an attribute, __capability binds to the struct.
However, when treated as a keyword, we effectively parse __capability as
a type qualifier on the pointer type.

The above case in particular is important as it occurs in the Morello
Linux kernel.

N.B. (like LLVM) we still use an attribute internally to represent
__capability to avoid too much churn outside of the parser. Since we no
longer want to accept the attribute from C/C++ code, though, we rename
the attribute to "cheri capability" (with a space) like other internal
GCC attributes.

We then adjust the testsuite accordingly.

While writing the patch for the C front end, it became apparent that we
have some tests that rely on __capability being accepted from C++,
despite us never really having added support for this. This happened to
work before since __capability got expanded to a GNU attribute, but the
C++ frontend was missing some diagnostics that we added for C, and
naturally had similar shortcomings due to relying on parsing
__capability as a GNU attribute.

Hence, we also update the C++ front end to handle __capability properly
as a cv-qualifier. For C++, we handle __capability in decl spec
sequences (i.e. in the deprecated prefix position) by making it a
first-class declaration specifier, and we handle the proper type
qualifier variant internally by plumbing through the attribute as a
C++11 standard attribute (which eventually gets applied in the
cdk_pointer case of the main loop in grokdeclarator). This proved to be
the least disruptive way of doing this.

N.B. for now we only support __capability on pointers in C++, in
particular we do not support the __capability keyword on reference
types. These are rejected. Since we don't currently support hybrid C++,
we have chosen to implement a minimal level of support such that
pure-capability code depending on the __capability keyword (e.g.
intrinsic wrappers and builtin tests) can be compiled. Adding support
for __capability references would require a wider adaptation of the C++
frontend code for hybrid CHERI.