Handle split stacks in unwinder
A signal handler can be registered with sigaction flags including SA_ONSTACK.
This flag indicates that the signal handler should be created on a different
stack to the current thread.

The mechanism by which the unwinder adjusts its stack pointer is by receiving a
delta between the current CFA and the target CFA, then adding that offset to the
stack pointer once the existing frame has been destroyed.
When the unwinder was called on one stack and needs to unwind to another stack
this addition of an offset gives us an invalid capability.

This patch resolves that by getting the unwinder to communicate to the epilogue
the new stack pointer that we want to end up with (rather than the offset
between the new stack pointer and the existing CFA).  The epilogue then uses
that value as the new stack pointer instead of adding an offset.

This means changing some generic code and the interface between generic code and
target code.  Such a change would break existing target code that does not get
updated.  While we have not been ensuring that other targets stay functional
with CHERI changes, we still do not want to knowingly break them.  Hence here we
choose to only use the new interface with purecap code.

I believe that changing the interface of this builtin is OK to do since it is
not documented and hence I believe it is only used by our runtime (which has
now been updated).

The adjustments that we make to the interface are:
  - __builtin_eh_return now takes an unwind_word type rather than the type for
    an offset_mode (ptr_mode) mode.
    (On non-purecap code this simply means a different integer -- usually of the
    same size, but sometimes larger.  On purecap code this means a uintcap_t).
  - For purecap only __builtin_eh_return now takes "the target value for csp"
    rather than "the offset between the current frames CFA and the target stack
    pointer".
  - For purecap only, the backend now receives "the target value for csp" rather
    than "the offset to be applied to csp after resetting to CFA" in
    EH_RETURN_STACKADJ_RTX.  (And hence this RTX needs to be in Pmode rather
    than POmode for purecap targets).

This fixes tests cleanup-10 and cleanup-11 in gcc.dg and g++.dg.  It also gets
nptl/tst-cancelx21 passing (on top of an earlier unwinder patch).

N.b. this patch leaves an unnecessary instruction in the epilogue of functions
which use `__builtin_eh_return`.  That is the instruction which moves the stack
pointer up to the CFA for the current function.  Before this patch that was
required since the unwinder provided the offset between the current functions
CFA and the target CFA so the epilogue needed to adjust the stack pointer both
by this offset and by the amount to get to the current functions CFA.
Now it is not necessary, but we leave it in for simplicity of this patch.